Bug fixed in setting the laser resolution in the GUI - fixed resolution variable from int to float

+++ direcs/src/gui.h
@@ -476,8 +476,8 @@ class Gui : public QMainWindow
 		QColor colorGraphicsSceneBackground;
 		int laserscannerAngleFront; /// this values holds the range of the laserscanner viewing angle. e.g. 180 oder 270 degrees.  just for reading the value here. @sa laserThread
 		int laserscannerAngleRear; /// this values holds the range of the laserscanner viewing angle. e.g. 180 oder 270 degrees.  just for reading the value here. @sa laserThread
-		int laserscannerResolutionFront; /// this values holds the resolution of the laserscanner. @sa laserThread
-		int laserscannerResolutionRear; /// this values holds the resolution of the laserscanner. @sa laserThread
+		float laserscannerResolutionFront; /// this values holds the resolution of the laserscanner. @sa laserThread
+		float laserscannerResolutionRear; /// this values holds the resolution of the laserscanner. @sa laserThread
 		int mLargestFreeAreaStartFront; /// this stores the free area angle from the front laser. Got it from @sa showLaserFrontAngles()
 		int mLargestFreeAreaEndFront;   /// this stores the free area angle from the front laser. Got it from @sa showLaserFrontAngles()
 		int mLargestFreeAreaStartRear;  /// this stores the free area angle from the rear  laser. Got it from @sa showLaserFrontAngles()
